holy squeak spirit of the code bass boosted
- do i have brain memory issue ? 
- did i get touched by the holy spirit of the saint code ? 
- i asked chatgpt to optimize the code, just ignored that stupid idiot 
- "Aweminus - riddim 4 dummies"
- patate dans le mentonade

+++ FileFormat/POG.cs
@@ -315,7 +315,7 @@ namespace hogs_gameEditor_wpf.FileFormat
                 team = (short)team,
                 objective = (short)objective,
                 ScriptGroup,
-                ScriptParameters = ScriptParameters.Select(x => (short)x).ToArray(),
+                ScriptParameters = ScriptParameters.Select(x => (byte)x).ToArray(),
                 fallback_position,
                 objectiveFlag = (short)objectiveFlag,
                 short1,
